fix: [UI] Split file handling into file-management.service
Fixes: #1070
